+++ app/src/test/java/org/simple/clinic/home/overdue/search/OverdueSearchUpdateTest.kt
@@ -63,6 +63,7 @@ class OverdueSearchUpdateTest {
 
   @Test
   fun `when overdue search results are loaded, then show overdue search results`() {
+    val searchQuery = "Babri"
     val facilityUuid = UUID.fromString("7dba16a0-1090-41f6-8e0c-0d97989de898")
     val overdueAppointments = listOf(TestData.overdueAppointment(
         facilityUuid = facilityUuid,
@@ -77,12 +78,12 @@ class OverdueSearchUpdateTest {
     val overdueSearchResult = PagingData.from(overdueAppointments)
 
     updateSpec
-        .given(defaultModel)
+        .given(defaultModel.overdueSearchQueryChanged(searchQuery))
         .whenEvent(OverdueSearchResultsLoaded(overdueSearchResult))
         .then(
             assertThatNext(
                 hasNoModel(),
-                hasEffects(ShowOverdueSearchResults(overdueSearchResult))
+                hasEffects(ShowOverdueSearchResults(overdueSearchResult, searchQuery))
             )
         )
   }