+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 - Don't show no patients found error view when search query is empty
 - Don't show empty search history items in overdue search screen
 - Prefill search query when search history item is clicked
+- Highlight search results with the matched search query
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-06-20-8300